+++ read_json.hpp
@@ -11,6 +11,6 @@ extern int geometry_within[GEOM_TYPES];
 extern int mb_geometry[GEOM_TYPES];
 
 void json_context(json_object *j);
-void parse_geometry(int t, json_object *j, drawvec &out, int op, const char *fname, int line, json_object *feature);
+void parse_geometry(int t, json_object *j, drawvec &out, int op, const char *fname, size_t line, json_object *feature);
 
-void stringify_value(json_object *value, int &type, std::string &stringified, const char *reading, int line, json_object *feature);
+void stringify_value(json_object *value, int &type, std::string &stringified, const char *reading, size_t line, json_object *feature);
